python-examples/hvac/access-tokens/requirements.txt
google-auth
google-cloud-storage
hvac
requests
http-constants